Examples in 1D cylindrical geometry (Jupyter)
Heat transfer examples (diffusion, with source term)

These notebooks are generally analytic vs. FVM solution, so may be used for testing
and benchmarking purposes. They also encode the correct formulae of the analytic
solutions, which is useful as these are sometimes reported in literature with typing
errors or ambiguities in the naming and units of the variables

The Jupyter Notebook ipynb format is not very convenient to maintain and update 
in a git context. However, notebooks are convenient for initial documentation
and demonstration purposes and for now, I will put these examples in notebooks. 
Other formats can be applied later on.